Isolate P2P sanitize utils
