+++ elements/lisk-p2p/src/utils/validate.ts
@@ -14,7 +14,6 @@
  */
 import { gte as isVersionGTE, valid as isValidVersion } from 'semver';
 import { isIP, isNumeric, isPort } from 'validator';
-import { constructPeerIdFromPeerInfo } from '.';
 import {
 	INCOMPATIBLE_NETWORK_REASON,
 	INCOMPATIBLE_PROTOCOL_VERSION_REASON,
@@ -29,7 +28,6 @@ import {
 	P2PDiscoveredPeerInfo,
 	P2PNodeInfo,
 	P2PPeerInfo,
-	PeerLists,
 	ProtocolMessagePacket,
 	ProtocolPeerInfo,
 	ProtocolRPCRequestPacket,
@@ -55,28 +53,6 @@ export const validatePeerAddress = (ip: string, wsPort: number): boolean => {
 	}
 
 	return true;
-};
-
-export const incomingPeerInfoSanitization = (
-	peerInfo: ProtocolPeerInfo,
-): P2PPeerInfo => {
-	const { ip, ...restOfPeerInfo } = peerInfo;
-
-	return {
-		ipAddress: ip,
-		...restOfPeerInfo,
-	};
-};
-
-export const outgoingPeerInfoSanitization = (
-	peerInfo: P2PPeerInfo,
-): ProtocolPeerInfo => {
-	const { ipAddress, ...restOfPeerInfo } = peerInfo;
-
-	return {
-		ip: ipAddress,
-		...restOfPeerInfo,
-	};
 };
 
 export const validatePeerInfoSchema = (rawPeerInfo: unknown): P2PPeerInfo => {
@@ -249,92 +225,5 @@ export const checkPeerCompatibility = (
 
 	return {
 		success: true,
-	};
-};
-
-export const sanitizePeerLists = (
-	lists: PeerLists,
-	nodeInfo: P2PPeerInfo,
-): PeerLists => {
-	const blacklistedPeers = lists.blacklistedPeers.filter(peerInfo => {
-		if (peerInfo.ipAddress === nodeInfo.ipAddress) {
-			return false;
-		}
-
-		return true;
-	});
-
-	const blacklistedIPs = blacklistedPeers.map(peerInfo => peerInfo.ipAddress);
-
-	const seedPeers = lists.seedPeers.filter(peerInfo => {
-		if (peerInfo.ipAddress === nodeInfo.ipAddress) {
-			return false;
-		}
-
-		if (blacklistedIPs.includes(peerInfo.ipAddress)) {
-			return false;
-		}
-
-		return true;
-	});
-
-	const fixedPeers = lists.fixedPeers.filter(peerInfo => {
-		if (peerInfo.ipAddress === nodeInfo.ipAddress) {
-			return false;
-		}
-
-		if (blacklistedIPs.includes(peerInfo.ipAddress)) {
-			return false;
-		}
-
-		return true;
-	});
-
-	const whitelisted = lists.whitelisted.filter(peerInfo => {
-		if (peerInfo.ipAddress === nodeInfo.ipAddress) {
-			return false;
-		}
-
-		if (blacklistedIPs.includes(peerInfo.ipAddress)) {
-			return false;
-		}
-
-		if (
-			fixedPeers
-				.map(constructPeerIdFromPeerInfo)
-				.includes(constructPeerIdFromPeerInfo(peerInfo))
-		) {
-			return false;
-		}
-
-		if (
-			seedPeers
-				.map(constructPeerIdFromPeerInfo)
-				.includes(constructPeerIdFromPeerInfo(peerInfo))
-		) {
-			return false;
-		}
-
-		return true;
-	});
-
-	const previousPeers = lists.previousPeers.filter(peerInfo => {
-		if (peerInfo.ipAddress === nodeInfo.ipAddress) {
-			return false;
-		}
-
-		if (blacklistedIPs.includes(peerInfo.ipAddress)) {
-			return false;
-		}
-
-		return true;
-	});
-
-	return {
-		blacklistedPeers,
-		seedPeers,
-		fixedPeers,
-		whitelisted,
-		previousPeers,
 	};
 };
